\name{calcIROLayer}
\alias{calcIROLayer}
\title{calcIROLayer}
\description{Calculate the isotopic ratio of the O-Layer.}
\usage{
calcIROLayer(somStockMineral, somStock, irS, irM)
}
\arguments{
  \item{somStockMineral}{numeric scalar: the CStock of the mineral soil (assumed not to change)}
  \item{somStock}{numeric vector: the CStock of the total soil}
  \item{irS}{isotopic ratio of the total SOM carbon}
  \item{irM}{isotopic ratio of the mineral SOM carbon (assumed not to change during this time)}
}

\value{numeric vector}

\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}





