\name{calcSteadyHcY_ICBM1}
\alias{calcSteadyHcY_ICBM1}
\title{calcSteadyHcY ICBM1}
\description{calculte cY and h from assuming steady state and decay constants}
\usage{
calcSteadyHcY_ICBM1(Ctot, iY, parms)
}
\arguments{
  \item{Ctot}{SOM C-Stock}
  \item{iY}{steady state input}
  \item{parms}{named numeric vector with entries kY,kO and cY and h}
}

\value{parms with entries cY and h replaced}

\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}





