\name{calcSteadyK_ICBM1}
\alias{calcSteadyK_ICBM1}
\title{calcSteadyK ICBM1}
\description{calculte decay constants from assuming steady state and remaining parameters}
\usage{
calcSteadyK_ICBM1(Ctot, cY, h, iY)
}
\arguments{
  \item{Ctot}{SOM C-Stock}
  \item{cY}{proportion of C-Stock in young pool}
  \item{h}{humification coefficient}
  \item{iY}{steady state input}
}

\value{named numeric vector with components kY and kO}

\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}





