\name{delta13Ccorrected14C}
\alias{delta13Ccorrected14C}
\title{delta13Ccorrected14C}
\description{correct 14C atomic ratio for fractionation by delta13C value }
\usage{
delta13Ccorrected14C(iR, delta13C = -25)
}
\arguments{
  \item{iR}{14C isotopic ratio, or fraction modern}
  \item{delta13C}{isotopic ratio of 13C in delta units for correction for fractionation}
}



\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}





