\name{DLLfuncTest}
\alias{DLLfuncTest}
\title{DLLfuncTest}
\description{Adjustment of \code{\link[deSolve]{DLLfunc}}, that does not coerce parms to double vector.}
\usage{
DLLfuncTest(func, times, y, parms, dllname, initfunc = dllname, rpar = NULL, ipar = NULL, nout = 0, outnames = NULL, forcings = NULL, initforc = NULL, fcontrol = NULL)
}
\arguments{
  \item{func}{
}
  \item{times}{
}
  \item{y}{
}
  \item{parms}{
}
  \item{dllname}{
}
  \item{initfunc}{
}
  \item{rpar}{
}
  \item{ipar}{
}
  \item{nout}{
}
  \item{outnames}{
}
  \item{forcings}{
}
  \item{initforc}{
}
  \item{fcontrol}{
}
}



\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}





