     \name{HowlandParameterPriors}
     \docType{data}
     \alias{HowlandParameterPriors}
     \title{Prior distribution of parameters for the Howland inversion study.}
     \description{     
     Prior distribution of parameters for the Howland inversion study.
     }
     \usage{HowlandParameterPriors}
     \format{component \code{parDistr} describes the distribution \describe{
		\item{trans}{kind of the distribution: normal (norm), lognormal (lognorm), or logitnormal (logitnorm) }
		\item{mu}{location parameter}
		\item{sigmaDiag}{scale parameter (note: not squared)}
		}
		Component \code{parms0} lists the mode for each parameter.
     }
     \source{Thomas Wutzler}
     \references{
     }
     \keyword{datasets}
