\name{ir14C2delta}
\alias{ir14C2delta}
\title{ir14C2delta}
\description{Convert atomic ratio to delta units.}
\usage{
ir14C2delta(iR, iR14CStandard = c14Constants$iR14CStandard)
}
\arguments{
  \item{iR}{numeric vector of atomic ratios}
  \item{iR14CStandard}{
}
}



\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}





