\name{origInput}
\alias{origInput}
\title{origInput}
\description{provide and unperturbed data series, neglect the standard deviation}
\usage{
origInput(input, padj)
}
\arguments{
  \item{input}{list of datastream matrices with first two columns time and value}
  \item{padj}{parameters}
}



\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}





