\name{plotHowlandFM}
\alias{plotHowlandFM}
\title{plotHowlandFM}
\description{Plot the fraction modern result of a model-data fit for the Howland data.}
\usage{
plotHowlandFM(res, obs)
}
\arguments{
  \item{res}{result of of.HowlandSteady}
  \item{obs}{observations}
}



\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}





