\name{initState.howland.ICBM1SteadyState}
\alias{initState.howland.ICBM1SteadyState}
\title{initState howland ICBM1SteadyState}
\description{Initial states from parms for ICBM1 in steady state}
\usage{
initState.howland.ICBM1SteadyState(padj, modMeta = modMetaICBM1(), fFmAtmosphere = fmAtmosphere)
}
\arguments{
  \item{padj}{must contain cY, kO, Ctot, yr0}
  \item{modMeta}{
}
  \item{fFmAtmosphere}{
}
}



\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}





